Move tina config for kitchen sink
