--- examples/kitchen-sink/.tina/.gitignore
@@ -1,2 +1,0 @@
-## For the monorepo this can be ignored
-__generated__
