+++ examples/kitchen-sink/lib/databaseConntection.ts
@@ -1,5 +1,5 @@
 import database from '../.tina/database'
-import { queries } from '../.tina/__generated__/types'
+import { queries } from '../tina/__generated__/types'
 import { resolve } from '@tinacms/datalayer'
 import type { TinaClient } from 'tinacms/dist/client'
