+++ examples/kitchen-sink/pages/collection/[collection]/[filename].tsx
@@ -1,6 +1,6 @@
 import { InferGetServerSidePropsType } from 'next'
 import { useTina } from 'tinacms/dist/react'
-import client from '../../../.tina/__generated__/client'
+import client from '../../../tina/__generated__/client'
 import { Json } from '../../../components/json'
 
 export default function Home(
