+++ examples/kitchen-sink/pages/node/[...filename].tsx
@@ -1,7 +1,7 @@
 import { InferGetServerSidePropsType } from 'next'
 import { Json } from '../../components/json'
 import { useTina } from 'tinacms/dist/react'
-import client from '../../.tina/__generated__/client'
+import client from '../../tina/__generated__/client'
 
 export default function Home(
   props: InferGetServerSidePropsType<typeof getServerSideProps>
