+++ examples/kitchen-sink/tina/config.tsx
